Extract MigrationWorkflow from InitWorkflows
InitWorkflows had grown into a dual-purpose class: the interactive init walkthrough plus the non-interactive migration helpers shared with InstallWorkflow. Pull the non-interactive pieces into a new internal static MigrationWorkflow in Commands.Shared:

  - MigrationSelection record

  - GetMigrationCandidates / BuildMigrationSelections

  - GetTrackedMigrationChannels / GetTrackedMigrationChannelName

  - ExecuteMigrationInPhases / BuildMigrationPhase{1,2}Requests

  - MergeInstallRequests / BuildMigrationInstallOptions

  - FilterRuntimeMigrationsAgainstDisk / RuntimeFolderExistsOnDisk

InitWorkflows keeps the interactive surface (PromptInstallsToMigrateIfDesired, PromptUserForMigration, FormatMigrationDisplayItems, the Phase 1/2 runner with banner+predownload, etc.). InstallWorkflow no longer has to reach into Init to find migration types.
